Rename baseline calibration binary as noise statistics
