+++ analysis/src/bin/alpha-g-wire-noise-statistics/plot.rs
