+++ analysis/src/bin/alpha-g-wire-noise-statistics/statistics/tests.rs
